+++ src/app/Components/ComponentsSamples/BarcodeScanning/BarcodeScanningSample.xaml.cs
@@ -21,9 +21,11 @@ public partial class BarcodeScanningSample
     {
         try
         {
-            await m_barcodeScanner.Start(CameraPreview, CameraFailed, settings =>
+            await m_barcodeScanner.Start(new BarcodeScannerStartOptions
             {
-                settings.OnValidBarcodeScanned = DidFindBarcode;
+                Preview = CameraPreview,
+                OnCameraFailed = CameraFailed,
+                OnBarcodeAcceptedAsync = HandleBarcodeAcceptedAsync
             });
         }
         catch (Exception exception)
@@ -38,23 +40,24 @@ public partial class BarcodeScanningSample
         App.Current.MainPage.DisplayAlert("Something failed!", e.Message, "Ok");
     }
 
-    private void DidFindBarcode(BarcodeScanResult barcodeScanResult)
+    private Task HandleBarcodeAcceptedAsync(BarcodeScanResult barcodeScanResult)
     {
         if (m_barCodeResultBottomSheet is
             {
                 HasBarCode: true
             })
         {
-            return;
+            return Task.CompletedTask;
         }
 
         m_barCodeResultBottomSheet = new BarcodeScanningResultBottomSheet();
         m_barCodeResultBottomSheet.Closed += BottomSheetClosed;
         m_barCodeResultBottomSheet.OpenWithBarCode(barcodeScanResult);
-        m_barcodeScanner.StopAndDispose();
+        m_barcodeScanner.PauseScanning();
+        return Task.CompletedTask;
     }
 
-    private async void BottomSheetClosed(object? sender, EventArgs e)
+    private void BottomSheetClosed(object? sender, EventArgs e)
     {
         if (m_barCodeResultBottomSheet != null)
         {
@@ -62,10 +65,10 @@ public partial class BarcodeScanningSample
         }
 
         m_barCodeResultBottomSheet = null;
-        _ = Start();
+        m_barcodeScanner.ResumeScanning();
     }
 
-    protected override async void OnAppearing()
+    protected override void OnAppearing()
     {
         _ = Start();
         base.OnAppearing();
@@ -78,6 +81,7 @@ public partial class BarcodeScanningSample
 
     private void Close(object? sender, EventArgs e)
     {
+        m_barcodeScanner.StopAndDispose();
         Shell.Current.Navigation.PopModalAsync();
     }
 }